#7120: refactor set for hearing / add to trial session modal helpers
