+++ web-client/src/presenter/computeds/addToTrialSessionModalHelper.js
@@ -1,9 +1,46 @@
-import { formatTrialSessionsForHelper } from './setForHearingModalHelper';
-import { sortBy } from 'lodash';
+import { isEmpty, isEqual, sortBy } from 'lodash';
 import { state } from 'cerebral';
 
-export const addToTrialSessionModalHelper = (get, applicationContext) => {
+const formatTrialSessionsForHelper = (trialSessions, applicationContext) => {
   const { SESSION_STATUS_GROUPS } = applicationContext.getConstants();
+
+  return trialSessions.map(trialSession => {
+    trialSession.startDateFormatted = applicationContext
+      .getUtilities()
+      .formatDateString(trialSession.startDate, 'MMDDYY');
+    switch (trialSession.sessionType) {
+      case 'Regular':
+      case 'Small':
+      case 'Hybrid':
+        trialSession.sessionTypeFormatted = trialSession.sessionType.charAt(0);
+        break;
+      case 'Special':
+        trialSession.sessionTypeFormatted = 'SP';
+        break;
+      case 'Motion/Hearing':
+        trialSession.sessionTypeFormatted = 'M/H';
+        break;
+    }
+    trialSession.optionText = `${trialSession.trialLocation} ${trialSession.startDateFormatted} (${trialSession.sessionTypeFormatted})`;
+
+    const allCases = trialSession.caseOrder || [];
+    const inactiveCases = allCases.filter(
+      sessionCase => sessionCase.removedFromTrial === true,
+    );
+
+    if (!isEmpty(allCases) && isEqual(allCases, inactiveCases)) {
+      trialSession.computedStatus = SESSION_STATUS_GROUPS.closed;
+    } else if (trialSession.isCalendared) {
+      trialSession.computedStatus = SESSION_STATUS_GROUPS.open;
+    } else {
+      trialSession.computedStatus = SESSION_STATUS_GROUPS.new;
+    }
+
+    return trialSession;
+  });
+};
+
+export const trialSessionsModalHelper = (get, applicationContext) => {
   const caseDetail = get(state.caseDetail);
   const { showAllLocations, trialSessionId, trialSessions } = get(state.modal);
 
@@ -18,10 +55,6 @@ export const addToTrialSessionModalHelper = (get, applicationContext) => {
     trialSessionsFormatted = formatTrialSessionsForHelper(
       trialSessionsFormatted,
       applicationContext,
-    ).filter(trialSession =>
-      [SESSION_STATUS_GROUPS.new, SESSION_STATUS_GROUPS.open].includes(
-        trialSession.computedStatus,
-      ),
     );
 
     if (showAllLocations) {
@@ -72,5 +105,26 @@ export const addToTrialSessionModalHelper = (get, applicationContext) => {
     trialSessionStatesSorted,
     trialSessionsFormatted,
     trialSessionsFormattedByState,
+  };
+};
+
+export const addToTrialSessionModalHelper = (get, applicationContext) => {
+  const { SESSION_STATUS_GROUPS } = applicationContext.getConstants();
+  let { trialSessionsFormatted, ...helperProps } = trialSessionsModalHelper(
+    get,
+    applicationContext,
+  );
+
+  if (trialSessionsFormatted) {
+    trialSessionsFormatted = trialSessionsFormatted.filter(trialSession =>
+      [SESSION_STATUS_GROUPS.new, SESSION_STATUS_GROUPS.open].includes(
+        trialSession.computedStatus,
+      ),
+    );
+  }
+
+  return {
+    ...helperProps,
+    trialSessionsFormatted,
   };
 };
